yaml: Rename several classes to put them under 'yaml' namespace.
To be consistent with json::document_tree and its friends.

* yaml_node_t -> yaml::node_t
* yaml_document_error -> yaml::document_error
* yaml_document_tree -> yaml::document_tree

+++ include/orcus/json_document_tree.hpp
@@ -49,7 +49,7 @@ public:
     virtual ~key_value_error() throw();
 };
 
-enum class node_t : int
+enum class node_t : uint8_t
 {
     /** node type is not set.  */
     unset = 0,
